old_python_setup/requirements.txt
pandas>=2.0.0
odfpy>=1.4.1
Pillow>=10.0.0
qrcode[pil]>=7.4.2

